feat: 내가 투표한 밸런스 게임 페이지
