+++ src/app/my/voted/page.tsx
@@ -1,0 +1,5 @@
+import VotedPage from '@/components/pages/my/voted/votedPage'
+
+export default function Voted() {
+  return <VotedPage />
+}
