+++ src/components/pages/my/voted/votedPage.tsx
@@ -1,0 +1,7 @@
+import HistoryPage from '@/components/pages/my/_shared/historyPage'
+
+const VotedPage = () => {
+  return <HistoryPage mode="voted" />
+}
+
+export default VotedPage
